+++ .travis_install_dependencies.sh
@@ -14,10 +14,10 @@ case "${DEPS:-}" in
 		;;
 	*)
 		if [ -n "${MIN_STABILITY:-}" ]; then
-			sed -i -e "s/\"minimum-stability\": \"stable\"/\"minimum-stability\": \"${MIN_STABILITY}\"/" composer.json
+			composer config minimum-stability "${MIN_STABILITY}"
 		fi
 
-		composer remove --no-update symfony/twig-bundle
+		composer remove --no-update symfony/config symfony/dependency-injection symfony/twig-bundle
 
 		if [ -n "${SYMFONY_VERSION:-}" ]; then
 			composer require --no-update --dev symfony/symfony:"${SYMFONY_VERSION}"
@@ -27,5 +27,9 @@ case "${DEPS:-}" in
 			composer require --no-update --dev twig/twig:"${TWIG_VERSION}"
 		fi
 esac
+
+if [ -n "${WITH_STATIC_ANALYSIS:-}" ]; then
+	composer require --no-update --dev phpstan/phpstan-shim
+fi
 
 composer update ${COMPOSER_UPDATE_ARGS:-}