+++ Tests/IntegrationTestBundle/Form/CommentFormType.php
@@ -3,11 +3,12 @@
 namespace Craue\TwigExtensionsBundle\Tests\IntegrationTestBundle\Form;
 
 use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 
 /**
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2017 Christian Raue
+ * @copyright 2011-2019 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class CommentFormType extends AbstractType {
@@ -16,17 +17,9 @@ class CommentFormType extends AbstractType {
 	 * {@inheritDoc}
 	 */
 	public function buildForm(FormBuilderInterface $builder, array $options) {
-		$useFqcn = method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
-
-		$builder->add('note', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\TextType' : 'text');
+		$builder->add('note', TextType::class);
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
-	public function getName() {
-		return $this->getBlockPrefix();
-	}
 	/**
 	 * {@inheritDoc}
 	 */