+++ Tests/Twig/Extension/FormExtensionTest.php
@@ -9,13 +9,13 @@ use PHPUnit\Framework\TestCase;
  * @group unit
  *
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2017 Christian Raue
+ * @copyright 2011-2019 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class FormExtensionTest extends TestCase {
 
 	/**
-	 * @var FormatNumberExtension
+	 * @var FormExtension
 	 */
 	protected $ext;