+++ Util/TwigFeatureDefinition.php
@@ -4,9 +4,10 @@ namespace Craue\TwigExtensionsBundle\Util;
 
 /**
  * for internal use only
+ * @internal
  *
  * @author Christian Raue <christian.raue@gmail.com>
- * @copyright 2011-2017 Christian Raue
+ * @copyright 2011-2019 Christian Raue
  * @license http://opensource.org/licenses/mit-license.php MIT License
  */
 class TwigFeatureDefinition {
@@ -16,7 +17,7 @@ class TwigFeatureDefinition {
 	public $alias;
 	public $options;
 
-	public function __construct($name, $methodName, $alias = null, $options = array()) {
+	public function __construct($name, $methodName, $alias = null, $options = []) {
 		$this->name = $name;
 		$this->methodName = $methodName;
 		$this->alias = $alias;